Reorg into module for kmer stuff
